BugFix: Fixed a NPE when selecting the shownotes tab before starting playback of an episode. It seems this caused the onCompletionListener to be called and it was trying to auto-delete an uninitialised episode with mPodcastName was null.

+++ src/org/sixgun/ponyexpress/service/PodcastPlayer.java
@@ -137,7 +137,8 @@ public class PodcastPlayer extends Service implements AudioManager.OnAudioFocusC
 				}
 				//Delete episode if preference is set.
 				final boolean delete_episode = mPrefs.getBoolean(getString(R.string.auto_delete_key), false);
-				if (delete_episode){
+				if (mPodcastName != null && delete_episode){
+					Log.d(TAG, "Trying to delete episode");
 					if (Utils.deleteFile(mPonyExpressApp, mRowID, mPodcastName)){
 						mPonyExpressApp.getDbHelper().update(mPodcastName, mRowID, EpisodeKeys.DOWNLOADED, "false");
 					}
